% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cluster_stability.r
\name{calc_cluster_stability}
\alias{calc_cluster_stability}
\title{Calculate cluster stability for each cluster}
\usage{
calc_cluster_stability(
  fs,
  flowsom_orig,
  scale,
  seed = NULL,
  boot,
  chnl,
  n_cluster
)
}
\arguments{
\item{fs}{flowSet.
Assumed appropriately transformed and compensated (but not necessary scaled).}

\item{flowsom_orig}{output from FlowSOM::FlowSOM, as applied to original data.}

\item{scale}{logical. If \code{TRUE}, the variables are centred and scaled.}

\item{seed}{integer. If not \code{NULL}, then the seed is set to this.}

\item{boot}{integer. Number of bootstrap samples to take.}

\item{chnl}{character vector. Channels to use for clustering.
Same as \code{colsToUse} for \code{FlowSOM::FlowSOM}.}

\item{n_cluster}{integer. Number of clusters.
Same meaning as \code{nClus} for \code{FlowSOM::FlowSOM}.}
}
\value{
A \code{tibble::tibble} with the following three
}
\description{
Calculate cluster stability for each cluster
}
