% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flowset_test.r
\name{.get_flowset_test}
\alias{.get_flowset_test}
\title{Get a FlowSet for demonstration purposes}
\usage{
.get_flowset_test()
}
\value{
A FlowSet of two samples.
}
\description{
Loads two FCS files from the \code{flowWorkspaceData} package.
}
